Folder re-organization with new 'ReadOnlyOrderedParquetDataset' and test cases all running.
