+++ tests/test_store/test_ordered_parquet_dataset/test_write/test_iter_merge_data.py
@@ -15,7 +15,7 @@ from numpy import array
 from pandas import DataFrame
 from pandas.testing import assert_frame_equal
 
-from oups.store.ordered_parquet_dataset.ordered_parquet_dataset import create_custom_opd
+from oups.store.ordered_parquet_dataset.ordered_parquet_dataset.base import create_custom_opd
 from oups.store.ordered_parquet_dataset.write.iter_merge_split_data import iter_merge_split_data
