+++ tests/test_store/test_ordered_parquet_dataset/test_write/test_write.py
@@ -12,7 +12,7 @@ from pandas import DataFrame
 from pandas import Timestamp
 
 from oups.defines import DIR_SEP
-from oups.store.ordered_parquet_dataset.ordered_parquet_dataset import create_custom_opd
+from oups.store.ordered_parquet_dataset.ordered_parquet_dataset.base import create_custom_opd
 from oups.store.ordered_parquet_dataset.write import write
